# api/models.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package api

// Item defines model for Item.
type Item struct {
	ID    *int     `json:"ID,omitempty"`
	Name  *string  `json:"name,omitempty"`
	Price *float32 `json:"price,omitempty"`
}

// CreateItemJSONRequestBody defines body for CreateItem for application/json ContentType.
type CreateItemJSONRequestBody = Item

// UpdateItemByIdJSONRequestBody defines body for UpdateItemById for application/json ContentType.
type UpdateItemByIdJSONRequestBody = Item